fix: audit remediation batch 2 - eliminate duplication (2.10.20)
Route 4 hand-rolled JSON cache implementations onto the shared utils/cache.py helpers (load_json_cache/save_json_cache), gaining cache hit/miss metrics on the huntarr/horizon caches for free. Hoist collect_tmdb_ids (duplicated verbatim in export_to_mdblist/export_to_simkl) to one module-level function. Hoist _calculate_rating_multiplier, _save_cache, and _print_similarity_breakdown off movie.py/tv.py onto BaseRecommender.
